% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectArabidopsisRoot.R
\docType{methods}
\name{canonicalizeName,TrenaProjectArabidopsisRoot-method}
\alias{canonicalizeName,TrenaProjectArabidopsisRoot-method}
\alias{canonicalizeName}
\title{map every possible gene name and/or symbol to a standard AT orf name}
\usage{
\S4method{canonicalizeName}{TrenaProjectArabidopsisRoot}(obj, name)
}
\arguments{
\item{obj}{An object of class TrenaProject}

\item{name}{A character string: gene symbol, orf name, odd alias}
}
\description{
map every possible gene name and/or symbol to a standard AT orf name
}
