% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectArabidopsisRoot.R
\docType{methods}
\name{findCandidateTranscriptionFactorsByMotifInSequence,TrenaProjectArabidopsisRoot-method}
\alias{findCandidateTranscriptionFactorsByMotifInSequence,TrenaProjectArabidopsisRoot-method}
\alias{findCandidateTranscriptionFactorsByMotifInSequence}
\title{identify transcription factors (using orf names) with motif match above threshold in the specified regions}
\usage{

  \S4method{findCandidateTranscriptionFactorsByMotifInSequence}{TrenaProjectArabidopsisRoot}(obj,
  tbl.regions, pwmMatchMinimumAsPercentage)
}
\arguments{
\item{obj}{An object of class TrenaProject}

\item{tbl.regions}{A data.frame with chrom, start and end columsn}

\item{pwmMatchMinimumAsPercentage}{a long integer (eg, 95L) expression minimum match threshold percentage}
}
\description{
uses the "Bioconductor motif matcher" aka Biostrings::matchPWM as wrapped by trena's MotifMatcher class
}
