% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectArabidopsisRoot.R
\docType{methods}
\name{getAllTranscriptionFactors,TrenaProjectArabidopsisRoot-method}
\alias{getAllTranscriptionFactors,TrenaProjectArabidopsisRoot-method}
\alias{getAllTranscriptionFactors}
\title{get all genes annotated by GO to}
\usage{
\S4method{getAllTranscriptionFactors}{TrenaProjectArabidopsisRoot}(obj,
  source)
}
\arguments{
\item{obj}{An object of class TrenaProject}
}
\description{
get all genes annotated by GO to
}
