% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectArabidopsisRoot.R
\docType{methods}
\name{getGeneNames,TrenaProjectArabidopsisRoot-method}
\alias{getGeneNames,TrenaProjectArabidopsisRoot-method}
\alias{getGeneNames}
\title{get orf and geneSymbol for orf or geneSymbol}
\usage{
\S4method{getGeneNames}{TrenaProjectArabidopsisRoot}(obj, name)
}
\arguments{
\item{obj}{An object of class TrenaProject}

\item{name}{An character string, either an orf or a gene symbol}
}
\description{
get orf and geneSymbol for orf or geneSymbol
}
