% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectArabidopsisRoot.R
\docType{class}
\name{TrenaProjectArabidopsisRoot-class}
\alias{TrenaProjectArabidopsisRoot-class}
\alias{.TrenaProjectArabidopsisRoot}
\alias{TrenaProjectArabidopsisRoot}
\title{TrenaProjectArabidopsisRoot-class}
\usage{
TrenaProjectArabidopsisRoot(quiet = TRUE)
}
\value{
An object of the TrenaProjectArabidopsisRoot class
}
\description{
Expression, variant and covariate data for the genes of interest (perhaps unbounded) for pre-term birth studies
}
